Fix race condition in mnesia_monitor
Mnesia_monitor detect mnesia down using a remote process link.
And net_kernel nodeup to detect that a node are reacable again.

If there is a short node communication problem.
The node-down and node-up events can happen before the remotely linked process generates an 'EXIT'.
When node-down and node-up events are recevied they are stored and later checked if the node came up just before mnesia flagged the node as down.