diameter: add support for sharing local port
Add possibility to share port for all connecting and listening transports over SCTP.
This is done by opening the socket in another process.
The process holding the socket will do the connect.
Once it is connected the association will be peeled off and handled by the dedicated process for that transport.