inet_drv: on one-to-one sctp sockets, buffer writes
As for TCP sockets the writes are put in a queue instead of returning {error, eagain}.
But a better approach might be to send {Port, ready_for_write} back to the process when the socket are ready for write again.
In that way you can queue stuff in erlang or decide what to do with the unwritten data up in Erlang.